+++ src/mako/gatekeeper/authorization/exceptions/AuthorizerException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\gatekeeper\authorization\exceptions;
+
+use mako\gatekeeper\exceptions\GatekeeperException;
+
+/**
+ * Authorizer exception.
+ */
+class AuthorizerException extends GatekeeperException
+{
+
+}